Update program to have a UI outside the console
index.html updated to include buttons for each of the moves, and elements
to display messages for the winner of each round as well as the entire
game. script.js refactored to work with UI components instead of console.
styles.css updated to make round winner message blue and game winner
message bold.